+++ packages/vue/src/components/organisms/SfHeader/SfHeader.vue
@@ -1,97 +1,126 @@
 <template>
   <div
     class="sf-header"
-    :class="{
-      'sf-header--has-mobile-search': hasMobileSearch,
-      'sf-header--has-mobile-navigation': hasMobileNavigation && isMobile,
-      'sf-header--is-sticky': isSticky,
-      'sf-header--is-hidden': !isVisible,
-    }"
+    :class="{ 'sf-header--is-sticky': sticky, 'sf-header--is-hidden': hidden }"
+    :style="stickyHeight"
   >
-    <div class="sf-header__sticky-container">
-      <header ref="header" class="sf-header__container">
-        <!--@slot Use this slot to replace logo with text or icon-->
+    <div class="sf-header__wrapper">
+      <header ref="header">
+        <!--@slot Use this slot to replace logo with text or image-->
         <slot name="logo" v-bind="{ logo, title }">
-          <SfImage
-            v-if="logo"
-            :src="logo"
-            :alt="title"
-            class="sf-header__logo"
-          />
-          <h1 v-else-if="title" class="sf-header__title">{{ title }}</h1>
-        </slot>
-        <nav
-          :class="{ 'desktop-only': !hasMobileNavigation }"
-          class="sf-header__navigation"
-        >
-          <!--@slot Use this slot to replace default navigation links -->
-          <slot name="navigation" />
-        </nav>
-        <!--@slot Use this slot to replace default search bar-->
-        <slot
-          name="search"
-          v-bind="{ searchPlaceholder, searchValue, hasMobileSearch }"
-        >
-          <SfSearchBar
-            :value="searchValue"
-            :placeholder="searchPlaceholder"
-            aria-label="Search"
-            class="sf-header__search"
-            :class="{ 'desktop-only': !hasMobileSearch }"
-            @input="$emit('change:search', $event)"
-            @enter="$emit('enter:search', $event)"
-          />
-        </slot>
-        <!--@slot Use this slot to replace default header icons with custom content-->
-        <slot
-          name="header-icons"
-          v-bind="{ accountIcon, wishlistIcon, cartIcon }"
-        >
-          <SfButton
-            v-for="icon in headerIcons"
-            :key="icon.name"
-            class="sf-button--pure sf-header__icon desktop-only"
-            :aria-label="icon.name"
-            :aria-pressed="activeIcon === icon.name ? 'true' : 'false'"
-            @click="$emit(`click:${icon.name}`)"
-          >
-            <SfIcon
-              :icon="icon.icon"
-              :has-badge="cartIsNotEmpty && icon.hasBadge === true"
-              :badge-label="cartItemsQty"
-              size="xs"
-              :class="{
-                'sf-header__icon--is-active': activeIcon === icon.name,
-              }"
+          <SfLink link="/">
+            <SfImage
+              v-if="logo"
+              :src="logo"
+              :alt="title"
+              class="sf-header__logo"
             />
-          </SfButton>
+            <h1 v-else class="sf-header__title">{{ title }}</h1>
+          </SfLink>
         </slot>
-        <!--@slot Use this slot to replace default header language selector on mobile -->
-        <slot name="language-selector" />
+        <div class="sf-header__aside">
+          <!-- @slot Use this slot for language or currency selector -->
+          <slot name="aside" />
+        </div>
+        <div class="sf-header__actions">
+          <nav class="sf-header__navigation">
+            <slot name="navigation"></slot>
+          </nav>
+          <!--@slot Use this slot to replace default search bar-->
+          <slot name="search" v-bind="{ searchValue, searchPlaceholder }">
+            <SfSearchBar
+              :value="searchValue"
+              :placeholder="searchPlaceholder"
+              aria-label="Search"
+              class="sf-header__search"
+              @input="$emit('change:search', $event)"
+              @enter="$emit('enter:search', $event)"
+            />
+          </slot>
+          <!--@slot Use this slot to replace default header icons with custom content-->
+          <slot
+            name="header-icons"
+            v-bind="{
+              activeIcon,
+              cartHasProducts,
+              cartItemsQty,
+              cartIcon,
+              wishlistIcon,
+              accountIcon,
+            }"
+          >
+            <div class="sf-header__icons">
+              <SfButton
+                v-if="accountIcon"
+                class="sf-button--pure sf-header__action"
+                @click="$emit('click:account')"
+              >
+                <SfIcon
+                  :icon="accountIcon"
+                  size="1.25rem"
+                  :class="{
+                    'sf-header__icon--is-active': activeIcon === 'account',
+                  }"
+                />
+              </SfButton>
+              <SfButton
+                v-if="wishlistIcon"
+                class="sf-button--pure sf-header__action"
+                @click="$emit('click:wishlist')"
+              >
+                <SfIcon
+                  :icon="wishlistIcon"
+                  size="1.25rem"
+                  :class="{
+                    'sf-header__icon--is-active': activeIcon === 'wishlist',
+                  }"
+                />
+              </SfButton>
+              <SfButton
+                v-if="cartIcon"
+                class="sf-button--pure sf-header__action"
+                @click="$emit('click:cart')"
+              >
+                <SfIcon
+                  class="sf-header__icon"
+                  :icon="cartIcon"
+                  :has-badge="cartHasProducts"
+                  :badge-label="cartItemsQty"
+                  size="1.25rem"
+                  :class="{
+                    'sf-header__icon--is-active': activeIcon === 'cart',
+                  }"
+                />
+              </SfButton>
+            </div>
+          </slot>
+        </div>
       </header>
     </div>
-    <div v-show="isSticky" class="sf-header__sticky-holder" :style="height" />
   </div>
 </template>
 <script>
 import Vue from "vue";
 import SfHeaderNavigationItem from "./_internal/SfHeaderNavigationItem.vue";
+Vue.component("SfHeaderNavigationItem", SfHeaderNavigationItem);
 import {
   mapMobileObserver,
   unMapMobileObserver,
 } from "../../../utilities/mobile-observer";
-Vue.component("SfHeaderNavigationItem", SfHeaderNavigationItem);
-import SfIcon from "../../atoms/SfIcon/SfIcon.vue";
+import { isClient } from "../../../utilities/helpers";
 import SfImage from "../../atoms/SfImage/SfImage.vue";
 import SfSearchBar from "../../molecules/SfSearchBar/SfSearchBar.vue";
 import SfButton from "../../atoms/SfButton/SfButton.vue";
+import SfIcon from "../../atoms/SfIcon/SfIcon.vue";
+import SfLink from "../../atoms/SfLink/SfLink.vue";
 export default {
   name: "SfHeader",
   components: {
-    SfIcon,
     SfImage,
     SfSearchBar,
     SfButton,
+    SfIcon,
+    SfLink,
   },
   props: {
     /**
@@ -140,24 +169,6 @@ export default {
       },
     },
     /**
-     * Header search on mobile
-     */
-    hasMobileSearch: {
-      type: Boolean,
-      default: false,
-    },
-    hasMobileNavigation: {
-      type: Boolean,
-      default: false,
-    },
-    /**
-     * Header sticky to top
-     */
-    isSticky: {
-      type: Boolean,
-      default: false,
-    },
-    /**
      * Header search placeholder
      */
     searchPlaceholder: {
@@ -175,81 +186,69 @@ export default {
      * Header cart items quantity
      */
     cartItemsQty: {
-      type: String,
+      type: [String, Number],
       default: "0",
     },
+    /**
+     * Header sticky to top
+     */
+    isSticky: {
+      type: Boolean,
+      default: false,
+    },
+    /**
+     * Header search on mobile
+     */
   },
   data() {
     return {
-      headerIcons: [
-        {
-          conditional: this.accountIcon,
-          icon: this.accountIcon,
-          name: "account",
-          hasBadge: false,
-        },
-        {
-          conditional: this.wishlistIcon,
-          icon: this.wishlistIcon,
-          name: "wishlist",
-          hasBadge: false,
-        },
-        {
-          conditional: this.cartIcon,
-          icon: this.cartIcon,
-          name: "cart",
-          hasBadge: true,
-        },
-      ],
-      isVisible: true,
-      isSearchVisible: true,
+      icons: [],
+      height: 0,
+      hidden: false,
       sticky: false,
-      scrollDirection: undefined, // false = down, true = up
+      scrollDirection: null,
       lastScrollPosition: 0,
-      animationStart: undefined,
-      animationLong: undefined,
+      animationStart: null,
+      animationLong: null,
       animationDuration: 300,
-      height: {},
     };
   },
   computed: {
     ...mapMobileObserver(),
-    cartIsNotEmpty() {
+    cartHasProducts() {
       return parseInt(this.cartItemsQty, 10) > 0;
+    },
+    stickyHeight() {
+      return {
+        "--_header-height": `${this.height}px`,
+      };
     },
   },
   watch: {
-    scrollDirection() {
-      if (typeof window === "undefined" || typeof document === "undefined")
-        return;
-      window.cancelAnimationFrame(this.animationLong);
-      this.animationLong = undefined;
-      this.animationStart = undefined;
-      this.animationLong = window.requestAnimationFrame(this.animationHandler);
+    scrollDirection: {
+      handler() {
+        if (!isClient) return;
+        window.cancelAnimationFrame(this.animationLong);
+        this.animationLong = null;
+        this.animationStart = null;
+        this.animationLong = window.requestAnimationFrame(
+          this.animationHandler
+        );
+      },
     },
     isMobile: {
       handler() {
-        if (typeof window === "undefined" || typeof document === "undefined")
-          return;
+        if (!isClient) return;
         this.$nextTick(() => {
-          const containerHeight = this.$refs.header;
-          this.height = {
-            height: `${containerHeight.offsetHeight}px`,
-          };
+          this.height = this.$refs.header.offsetHeight;
         });
       },
       immediate: true,
     },
-    hasMobileSearch: {
-      handler() {
-        this.$nextTick(() => {
-          if (typeof window === "undefined" || typeof document === "undefined")
-            return;
-          const computedContainer = window.getComputedStyle(this.$refs.header);
-          this.height = {
-            height: computedContainer.height,
-          };
-        });
+    isSticky: {
+      handler(isSticky) {
+        if (!isClient) return;
+        this.sticky = isSticky;
       },
       immediate: true,
     },
@@ -260,10 +259,12 @@ export default {
     }
   },
   beforeDestroy() {
-    if (this.isSticky) {
-      window.removeEventListener("scroll", this.scrollHandler);
-    }
     unMapMobileObserver();
+    if (this.isSticky) {
+      window.removeEventListener("scroll", this.scrollHandler, {
+        passive: true,
+      });
+    }
   },
   methods: {
     animationHandler(timestamp) {
@@ -275,15 +276,16 @@ export default {
         );
         return;
       }
-      this.isVisible = this.scrollDirection;
+      this.hidden = this.scrollDirection === "down";
     },
     scrollHandler() {
-      if (typeof window === "undefined" || typeof document === "undefined")
-        return;
+      if (!isClient) return;
       const currentScrollPosition =
         window.pageYOffset || document.documentElement.scrollTop;
-      this.scrollDirection = currentScrollPosition < this.lastScrollPosition;
-      this.isSearchVisible = currentScrollPosition <= 0;
+      if (currentScrollPosition >= this.height) {
+        this.scrollDirection =
+          currentScrollPosition < this.lastScrollPosition ? "up" : "down";
+      }
       this.lastScrollPosition = currentScrollPosition;
     },
   },